% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speechanalysis-package.R
\docType{package}
\name{speechanalysis-package}
\alias{speechanalysis}
\alias{speechanalysis-package}
\title{speechanalysis: A Package to Provide Graphical Summaries of Speech Data}
\description{
Provides a number of functions which produce graphical summaries of speech data. An example dataset is provided, 
    which consists of 10 speeches made by Donald Trump throughout September 2020.
}
\author{
\strong{Maintainer}: Alex Mounsey \email{alex.mounsey@postgrad.plymouth.ac.uk}

}
\keyword{internal}
