% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trump_speeches.R
\docType{data}
\name{trump_speeches}
\alias{trump_speeches}
\title{Trump Speech Data}
\format{
A data.frame with 10 rows and 3 columns:
\describe{
\item{speech}{Speech transcript}
\item{location}{Location of speech}
\item{date}{Date of speech}
}
}
\source{
University of Plymouth, MATH513 Assignment 2020/21
}
\usage{
trump_speeches
}
\description{
A dataset consisting of 10 speeches made by Donald Trump throughout September 2020.
}
\examples{
plot_tf_idf(trump_speeches, 10)
word_frequency(trump_speeches, c("america", "biden", "China"))
zipfs_law(trump_speeches, 1)

}
\keyword{datasets}
