% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_frequency.R
\name{word_frequency}
\alias{word_frequency}
\title{Graph of word usage of a data set}
\usage{
word_frequency(df, words_to_graph)
}
\arguments{
\item{df}{A dataframe with the following columns: 'speech' <chr>, 'location' <chr>, 'date' <date>}

\item{words_to_graph}{A collection of words for which the user wants graphs plotted}
}
\value{
A named vector of a plot:
\describe{
\item{plot}{The data plotted in a graph.}
}
}
\description{
This function provides a graph using ggplt2 of a data set containing speeches.
}
\examples{
word_frequency(trump_speeches, c("america", "biden", "China"))

}
\author{
10701983 \email{10701983}
}
