% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speechanalysis-package.R
\docType{package}
\name{speechanalysis-package}
\alias{speechanalysis}
\alias{speechanalysis-package}
\title{speechanalysis: A Package to Provide Graphical Summaries of Speech Data}
\description{
Provides a number of functions which produce graphical summaries of speech data. An example dataset is provided, 
    which consists of 10 speeches made by Donald Trump throughout September 2020.
}
\author{
\strong{Maintainer}: 10696253 10696253 \email{10696253@postgrad.plymouth.ac.uk}

Authors:
\itemize{
  \item 10701983 10701983 \email{10701983@postgrad.plymouth.ac.uk}
  \item 10570155 10570155 \email{10570155@postgrad.plymouth.ac.uk}
}

}
\keyword{internal}
